+++ default.nix
@@ -1,13 +1,13 @@
 let
   holonixPath = (import ./nix/sources.nix).holonix;
   holonix = import (holonixPath) {
-    holochainVersionId = "v0_0_161";
+    holochainVersionId = "v0_0_170";
   };
   nixpkgs = holonix.pkgs;
 in
 nixpkgs.mkShell {
   inputsFrom = [ holonix.main ];
   packages = with nixpkgs; [
-    nodejs-16_x
+    nodejs
   ];
 }