+++ dnas/clutter/integrity_zomes/mews/src/lib.rs
@@ -17,12 +17,28 @@ pub enum MewType {
 }
 
 #[hdk_entry_helper]
+#[derive(Clone)]
+pub enum LinkTarget {
+    Mention(AgentPubKey),
+    URL(String),
+}
+
+#[derive(Debug, Serialize, Deserialize, SerializedBytes)]
 #[serde(rename_all = "camelCase")]
+pub struct CreateMewInput {
+    pub mew_type: MewType,
+    pub text: Option<String>,
+    pub links: Option<Vec<LinkTarget>>,
+}
+
+#[hdk_entry_helper]
+#[serde(rename_all = "camelCase")]
+#[derive(Clone)]
 pub struct MewContent {
     pub text: String,
     // "Visit this web site ^link by @user about #hashtag to earn $cashtag! Also read this humm earth post ^link (as an HRL link)"
     // [^links in the mewstring in sequence]
-    // links: Vec<LinkTypes>,
+    pub links: Option<Vec<LinkTarget>>,
     // mew_images: Vec<EntryHash>, //Vec of image links hashes to retrieve
 }
 
@@ -69,3 +85,28 @@ pub const LICK_PATH_SEGMENT: &str = "lick";
 pub const REPLY_PATH_SEGMENT: &str = "reply";
 pub const MEWMEW_PATH_SEGMENT: &str = "mewmew";
 pub const QUOTE_PATH_SEGMENT: &str = "quote";
+
+#[hdk_extern]
+pub fn validate(op: Op) -> ExternResult<ValidateCallbackResult> {
+    match op.to_type::<EntryTypes, LinkTypes>().unwrap() {
+        OpType::StoreEntry(entry) => match entry {
+            OpEntry::CreateEntry { entry_type, .. } => match entry_type {
+                EntryTypes::Mew(mew) => match mew.content {
+                    None => Ok(ValidateCallbackResult::Valid),
+                    Some(mew_content) => {
+                        if mew_content.text.len() <= 200 {
+                            Ok(ValidateCallbackResult::Valid)
+                        } else {
+                            Ok(ValidateCallbackResult::Invalid(
+                                "mew must not be longer than 200 characters".to_string(),
+                            ))
+                        }
+                    }
+                },
+                EntryTypes::MewContent(_) => Ok(ValidateCallbackResult::Valid),
+            },
+            _ => Ok(ValidateCallbackResult::Valid),
+        },
+        _ => Ok(ValidateCallbackResult::Valid),
+    }
+}