+++ ui/src/pages/TagMewsFeed.vue
@@ -13,10 +13,12 @@ import {
   getMewsWithMention,
 } from "@/services/clutter-dna";
 import { onMounted, computed, ref, watch } from "vue";
-import { FeedMew, TAG_SYMBOLS } from "@/types/types";
+import { FeedMew } from "@/types/types";
 import { showError } from "@/utils/notification";
 import { useRouter } from "vue-router";
 import MewList from "../components/MewList.vue";
+import { deserializeHash } from "@holochain-open-dev/utils";
+import { TAG_SYMBOLS } from "@/utils/tags";
 
 const router = useRouter();
 const currentRoute = computed(() => router.currentRoute.value);
@@ -25,6 +27,11 @@ const tag = computed(() =>
   Array.isArray(currentRoute.value.params.tag)
     ? currentRoute.value.params.tag[0]
     : currentRoute.value.params.tag
+);
+const agentPubKey = computed(() =>
+  Array.isArray(currentRoute.value.query.agentPubKey)
+    ? currentRoute.value.query.agentPubKey[0]
+    : currentRoute.value.query.agentPubKey
 );
 
 const loading = ref(false);
@@ -38,7 +45,9 @@ const loadMewsFeed = async () => {
     } else if (tagSymbol.value === TAG_SYMBOLS.HASHTAG) {
       mews.value = await getMewsWithHashtag(tagSymbol.value + tag.value);
     } else if (tagSymbol.value === TAG_SYMBOLS.MENTION) {
-      mews.value = await getMewsWithMention(tagSymbol.value + tag.value);
+      mews.value = await getMewsWithMention(
+        deserializeHash(agentPubKey.value || "")
+      );
     }
   } catch (error) {
     showError(error);